Added restaurant page, navigation to restaurant page from map and dish pages, and created a loading animation
