+++ app.json
@@ -9,7 +9,7 @@
     "orientation": "portrait",
     "icon": "./assets/icon.png",
     "splash": {
-      "image": "./assets/splash.png",
+      "image": "./styles/loading.json",
       "resizeMode": "contain",
       "backgroundColor": "#ffffff"
     },
