+++ package.json
@@ -22,9 +22,12 @@
     "apollo-upload-client": "^14.1.3",
     "axios": "^0.21.1",
     "expo": "~40.0.0",
+    "expo-app-loading": "^1.0.3",
     "expo-status-bar": "~1.0.3",
     "firebase": "^8.2.6",
     "graphql": "^15.5.0",
+    "lottie-ios": "^3.2.0",
+    "lottie-react-native": "~2.6.1",
     "mongoose": "^5.12.4",
     "react": "16.13.1",
     "react-dom": "16.13.1",
